Fix #121.
Remove the use of the remove_user_from_blog hook because it used our delete_data function incorrectly. And we had made incorrect assumptions of what the remove_user_from_blog hook was doing.

It is not necessary for simple:press and, because of the way that hook works, is not really recommended for use anywhere else - doing too many thing in it can cause php timeouts which could cause a subsite delete to fail.